requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
Flask-RESTful